Move resin atmos cleaning system out of SmokeSystem
