+++ Content.Shared/Chemistry/Components/SmokeComponent.cs
@@ -37,11 +37,4 @@ public sealed partial class SmokeComponent : Component
     /// </summary>
     [DataField, ViewVariables(VVAccess.ReadWrite), AutoNetworkedField]
     public float Duration = 10;
-
-    /// <summary>
-    /// Assmos - Extinguisher Nozzle. 
-    /// </summary>
-    [DataField]
-    public bool IsResin = false;
-
 }
