+++ Content.Shared/Explosion/Components/OnTrigger/SmokeOnTriggerComponent.cs
@@ -39,10 +39,4 @@ public sealed partial class SmokeOnTriggerComponent : Component
     /// </remarks>
     [DataField, ViewVariables(VVAccess.ReadWrite)]
     public Solution Solution = new();
-
-    /// <remarks>
-    /// Assmos - Extinguisher Nozzle.
-    /// </remarks>
-    [DataField]
-    public bool IsResin = false;
 }
